lib/requirements.txt
scons>=4
fastparquet>=0.4.0
pandas>=1.0.5
numpy>=1.19.1
matplotlib>=3.3.0
scipy>=1.5.2
PyYAML>=5.4.1
git+https://github.com/mcaceresb/tablefill
